This is a function that searches for a value in a sorted array of integers using the Interpolation search algorithm

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -24,7 +24,7 @@ int interpolation_search(int *array, size_t size, int value)
 
 	while (low <= high && value >= array[low] && value <= array[high])
 	{
-		pos = low + (((double)high - low) / (array[high] - array[low])) *
+		pos = low + ((high - low) / (array[high] - array[low])) *
 			(value - array[low]);
 		printf("Value checked array[%lu] = [%d]\n", pos, array[pos]);
